format code

+++ src/models/ClusterGCN.py
@@ -12,7 +12,6 @@ import time
 from tqdm import tqdm
 
 
-
 class ClusterGCN(GraphSamplingBase):
     # Implemented base on https://github.com/rusty1s/pytorch_geometric/blob/master/examples/graph_saint.py
     def __init__(self, args, data, train_idx, processed_dir):
@@ -25,28 +24,27 @@ class ClusterGCN(GraphSamplingBase):
         else:
             base_gnnconv = SAGEConv
 
-        ## build model
+        # build model
         self.convs = torch.nn.ModuleList()
         self.convs.append(base_gnnconv(self.num_feats, self.dim_hidden))
         for _ in range(self.num_layers - 2):
             self.convs.append(base_gnnconv(self.dim_hidden, self.dim_hidden))
         self.convs.append(base_gnnconv(self.dim_hidden, self.num_classes))
 
-
-        ## load data
-        sample_size = max(1, int(args.batch_size / (data.num_nodes / args.num_parts)))
-        cluster_data = ClusterData(data, num_parts=args.num_parts, recursive=False, save_dir=self.save_dir)
-        self.train_loader = ClusterLoader(cluster_data, batch_size=sample_size, shuffle=True, num_workers=0)
+        # load data
+        sample_size = max(
+            1, int(args.batch_size / (data.num_nodes / args.num_parts)))
+        cluster_data = ClusterData(
+            data, num_parts=args.num_parts, recursive=False, save_dir=self.save_dir)
+        self.train_loader = ClusterLoader(
+            cluster_data, batch_size=sample_size, shuffle=True, num_workers=0)
 
         self.saved_args = vars(args)
         self.reset_parameters()
 
-
     def reset_parameters(self):
         for conv in self.convs:
             conv.reset_parameters()
-
-
 
     def forward(self, x, edge_index):
         for i, conv in enumerate(self.convs):
@@ -56,18 +54,14 @@ class ClusterGCN(GraphSamplingBase):
                 x = F.dropout(x, p=self.dropout, training=self.training)
         return x
 
-
-
-
     def train_net(self, input_dict):
 
         device = input_dict["device"]
         optimizer = input_dict["optimizer"]
         loss_op = input_dict["loss_op"]
 
-        
         total_loss = total_correct = 0
-        for batch in tqdm( self.train_loader ):
+        for batch in tqdm(self.train_loader):
             batch = batch.to(device)
             if batch.train_mask.sum() == 0:
                 continue
@@ -75,10 +69,12 @@ class ClusterGCN(GraphSamplingBase):
             out = self(batch.x, batch.edge_index)
             if isinstance(loss_op, torch.nn.NLLLoss):
                 out = F.log_softmax(out, dim=-1)
-                loss = loss_op(out[batch.train_mask], batch.y[batch.train_mask])
+                loss = loss_op(out[batch.train_mask],
+                               batch.y[batch.train_mask])
             else:
                 loss = loss_op(
-                    out[batch.train_mask], batch.y[batch.train_mask].type_as(out)
+                    out[batch.train_mask], batch.y[batch.train_mask].type_as(
+                        out)
                 )
             loss.backward()
             optimizer.step()
@@ -88,5 +84,6 @@ class ClusterGCN(GraphSamplingBase):
             else:
                 total_correct += int(out.eq(batch.y).sum())
 
-        train_size = self.train_size if isinstance(loss_op, torch.nn.NLLLoss) else self.train_size * self.num_classes
+        train_size = self.train_size if isinstance(
+            loss_op, torch.nn.NLLLoss) else self.train_size * self.num_classes
         return total_loss / len(self.train_loader), total_correct / train_size
