+++ src/models/GraphSAINT.py
@@ -19,7 +19,6 @@ from tqdm import tqdm
 from .base import GraphSamplingBase
 
 
-
 class GraphSAINT(GraphSamplingBase):
     # Implemented base on https://github.com/rusty1s/pytorch_geometric/blob/master/examples/graph_saint.py
     def __init__(self, args, data, train_idx, processed_dir):
@@ -28,13 +27,12 @@ class GraphSAINT(GraphSamplingBase):
         self.dropout = args.dropout
         self.args = args
 
-
         if args.gnn_type == "gnn":
             # base_gnnconv = GraphConv
             base_gnnconv = SAGEConv
         elif args.gnn_type == "mlp":
             # base_gnnconv = GraphConv_MLP
-            base_gnnconv =  SAGEConv_MLP
+            base_gnnconv = SAGEConv_MLP
         else:
             # base_gnnconv = GraphConv
             base_gnnconv = SAGEConv
@@ -47,39 +45,34 @@ class GraphSAINT(GraphSamplingBase):
         self.convs.append(base_gnnconv(self.dim_hidden, self.num_classes))
         # self.lin = Linear(self.num_layers * self.dim_hidden, self.num_classes)
 
-
-        ## data load
+        # data load
         row, col = self.edge_index = data.edge_index
         data.edge_weight = 1.0 / degree(col, data.num_nodes)[col]
 
-
         self.train_loader = RWSampler(
-                                    data, 
-                                    self.batch_size, 
-                                    args.walk_length, 
-                                    num_steps=args.num_steps, 
-                                    save_dir=self.save_dir, 
-                                    sample_coverage=args.sample_coverage,
-                                    num_workers=0
-                                    )
+            data,
+            self.batch_size,
+            args.walk_length,
+            num_steps=args.num_steps,
+            save_dir=self.save_dir,
+            sample_coverage=args.sample_coverage,
+            num_workers=0
+        )
 
-        ##reset_parameters
+        # reset_parameters
         self.reset_parameters()
         self.saved_args = vars(args)
-        
+
         aggr = "add" if self.use_norm else "mean"
         self.set_aggr(aggr)
-
 
     def reset_parameters(self):
         for conv in self.convs:
             conv.reset_parameters()
 
-
     def set_aggr(self, aggr):
         for conv in self.convs:
             conv.aggr = aggr
-
 
     def forward(self, x, edge_index, edge_weight=None):
         for i, conv in enumerate(self.convs):
@@ -89,17 +82,14 @@ class GraphSAINT(GraphSamplingBase):
                 x = F.dropout(x, p=self.dropout, training=self.training)
         return x
 
-
-
     def train_net(self, input_dict):
 
         device = input_dict["device"]
         optimizer = input_dict["optimizer"]
         loss_op = input_dict["loss_op"]
 
-
         total_loss = total_correct = 0
-        for batch in tqdm( self.train_loader ):
+        for batch in tqdm(self.train_loader):
             batch = batch.to(device)
             if batch.train_mask.sum() == 0:
                 continue
@@ -120,10 +110,12 @@ class GraphSAINT(GraphSamplingBase):
                     loss = F.nll_loss(out, batch.y, reduction="none")
                     loss = (loss * batch.node_norm)[batch.train_mask].sum()
                 else:
-                    loss = loss_op(out[batch.train_mask], batch.y[batch.train_mask])
+                    loss = loss_op(out[batch.train_mask],
+                                   batch.y[batch.train_mask])
             else:
                 loss = loss_op(
-                    out[batch.train_mask], batch.y[batch.train_mask].type_as(out)
+                    out[batch.train_mask], batch.y[batch.train_mask].type_as(
+                        out)
                 )
 
             loss.backward()
@@ -141,8 +133,5 @@ class GraphSAINT(GraphSamplingBase):
             if isinstance(loss_op, torch.nn.NLLLoss)
             else self.train_size * self.num_classes
         )
-
-
-
 
         return total_loss / len(self.train_loader), total_correct / train_size
